Update DOM values when attributes are changed in setter
`this` is not bound to the Proxy in a class setter, so changing values
other than the original property does not go through the Proxy's `set`
(calling `updateDOM`).

Because of this, when a property has changed, compare old and new value
for every `data-bind` reference, and update DOM if necessary.